Revert "Fish AI rewrite"
