+++ Assets/Prefabs/FishPrefabs/BusFish.prefab
@@ -205,6 +205,8 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   moveSpeed: -5
   rb: {fileID: 1999087896956115757}
+  following: 0
+  distance: 0
   damage: 20
   damageTaken: 25
   seaLineObject: {fileID: 0}
@@ -221,7 +223,7 @@ Rigidbody2D:
   m_Simulated: 1
   m_UseFullKinematicContacts: 0
   m_UseAutoMass: 0
-  m_Mass: 1000000
+  m_Mass: 5
   m_LinearDrag: 0
   m_AngularDrag: 0.05
   m_GravityScale: 0
@@ -284,7 +286,7 @@ CapsuleCollider2D:
   m_CallbackLayers:
     serializedVersion: 2
     m_Bits: 4294967295
-  m_IsTrigger: 1
+  m_IsTrigger: 0
   m_UsedByEffector: 0
   m_UsedByComposite: 0
   m_Offset: {x: -0.5, y: 0}
