--- Assets/Scripts/Fish/OLDFishMovement.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: bc39f415be494474fba29a418cbcdd6a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
